Move minimap and bomb counter to `gui` directory
